Wed, Jul 3, 2024, 2:02 PM +05:30

+++ world/rough2.c
@@ -12,8 +12,8 @@ int main(void)
     }
     while (card_number <= 0);
 
-    long quotient = card_number/3;
+    int quotient = card_number/3;
 
-    printf("%li\n", quotient);
+    printf("%i\n", quotient);
 
 }
